third_party/blink/web_tests/external/wpt/app-history/navigate-event/navigate-anchor-same-origin-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<a id="a" href="foo.html"></a>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_true(e.cancelable);
    assert_true(e.canRespond);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.formData, null);
    assert_equals(new URL(e.destination.url).pathname,
                  "/app-history/navigate-event/foo.html");
    assert_false(e.destination.sameDocument);
    e.preventDefault();
  });
  a.click();
}, "<a> cross-document (but same-origin) navigate event");
</script>
